Memindahkan atribute kedalam kelas user
1. Memindahkan atribute firstname, lastname, address, dan gender kedalam kelas user

2. Menghapus parameter firstname, lastname, address, dan gender  pada constructor

3. Membuat contructor dan getter setter pada kelas user
